refactored/updated import code to allow for file:// protocol in urls
