+++ controllers/collections.py
@@ -1,5 +1,5 @@
 import json
-import urllib2
+from urllib2 import HTTPError
 
 from bson import json_util
 import cherrypy
@@ -7,7 +7,7 @@ from pandas import read_csv
 
 from config.db import db
 from lib.constants import BAMBOO_ID, SOURCE
-from lib.utils import df_to_mongo, mongo_decode_keys, df_to_hexdigest
+from lib.utils import df_to_mongo, mongo_decode_keys, df_to_hexdigest, open_data_file
 import models.collection
 
 class Collections(object):
@@ -49,34 +49,21 @@ class Collections(object):
         Read data from URL 'url'.
         If URL is not provided and data is provided, read posted data 'data'.
         '''
-        def open_data_file(url):
-            import re
-            open_url = lambda url: urllib2.urlopen(url)
-            protocols = {
-                'http':  open_url,
-                'https': open_url,
-                'file':  lambda path: path,
-            }
-            regex = re.compile(
-                '^(?P<url>(?P<protocol>%s):\/\/(?P<path>.+))$' \
-                % '|'.join(protocols)
-            )
-            match = re.match(regex, url)
-            if match:
-                protocols[match.groupdict()['protocol']](url)
-            return None
         f = open_data_file(url)
-        if f:
+        if not f: return # could not get a file handle
+        try:
             df = read_csv(f, na_values=['n/a'])
-            digest = df_to_hexdigest(df)
-            num_rows_with_digest = db().collections.find(
-                    {BAMBOO_ID: digest}).count()
-            if not num_rows_with_digest:
-                df = df_to_mongo(df)
-                # add metadata to file
-                for e in df:
-                    e[BAMBOO_ID] = digest
-                    e[SOURCE] = url
-                # insert data into collections
-                db().collections.insert(df)
-            return json.dumps({'id': digest})
+        except (IOError, HTTPError):
+            return # error reading file/url
+        digest = df_to_hexdigest(df)
+        num_rows_with_digest = db().collections.find(
+                {BAMBOO_ID: digest}).count()
+        if not num_rows_with_digest:
+            df = df_to_mongo(df)
+            # add metadata to file
+            for e in df:
+                e[BAMBOO_ID] = digest
+                e[SOURCE] = url
+            # insert data into collections
+            db().collections.insert(df)
+        return json.dumps({'id': digest})
